separated logic from components
