+++ todolist/src/app/accessoryClasses/logic/PriorityLogic.ts
@@ -1,0 +1,3 @@
+/**
+ * Created by Павел on 05-Mar-17.
+ */
